+++ engine/apps/alerts/tests/test_check_escalation_finished_task.py
@@ -9,11 +9,13 @@ from apps.alerts.models import EscalationPolicy
 from apps.alerts.tasks.check_escalation_finished import (
     AlertGroupEscalationPolicyExecutionAuditException,
     audit_alert_group_escalation,
+    check_alert_group_personal_notifications_task,
     check_escalation_finished_task,
     check_personal_notifications_task,
     send_alert_group_escalation_auditor_task_heartbeat,
 )
 from apps.base.models import UserNotificationPolicy, UserNotificationPolicyLogRecord
+from apps.twilioapp.models import TwilioSMS, TwilioSMSstatuses
 
 MOCKED_HEARTBEAT_URL = "https://hello.com/lsdjjkf"
 
@@ -406,6 +408,7 @@ def test_check_escalation_finished_task_calls_audit_alert_group_personal_notific
     make_alert_receive_channel,
     make_alert_group_that_started_at_specific_date,
     make_user_notification_policy_log_record,
+    make_sms_record,
     caplog,
 ):
     organization, user = make_organization_and_user()
@@ -460,7 +463,7 @@ def test_check_escalation_finished_task_calls_audit_alert_group_personal_notific
     # records created > 5 mins ago
     alert_group1.personal_log_records.update(created_at=now - timezone.timedelta(minutes=7))
 
-    # alert_group2: notify user, notification failed
+    # alert_group2: notify user, notification failed; triggered sms, sent status
     make_user_notification_policy_log_record(
         author=user,
         alert_group=alert_group2,
@@ -475,8 +478,24 @@ def test_check_escalation_finished_task_calls_audit_alert_group_personal_notific
         notification_step=UserNotificationPolicy.Step.NOTIFY,
         type=UserNotificationPolicyLogRecord.TYPE_PERSONAL_NOTIFICATION_FAILED,
     )
+    make_user_notification_policy_log_record(
+        author=user,
+        alert_group=alert_group2,
+        notification_policy=user_notification_policy,
+        notification_step=UserNotificationPolicy.Step.NOTIFY,
+        type=UserNotificationPolicyLogRecord.TYPE_PERSONAL_NOTIFICATION_TRIGGERED,
+    )
+    # no failed or succeed record, but SMS was sent (without Twilio delivered confirmation yet)
+    sms_record = make_sms_record(
+        receiver=user,
+        represents_alert_group=alert_group2,
+        notification_policy=user_notification_policy,
+    )
+    sent_sms = TwilioSMS.objects.create(sid="someid", sms_record=sms_record, status=TwilioSMSstatuses.SENT)
     # records created > 5 mins ago
     alert_group2.personal_log_records.update(created_at=now - timezone.timedelta(minutes=7))
+    sent_sms.created_at = now - timezone.timedelta(minutes=6)
+    sent_sms.save()
 
     # alert_group3: notify user, missing completion
     make_user_notification_policy_log_record(
@@ -502,15 +521,22 @@ def test_check_escalation_finished_task_calls_audit_alert_group_personal_notific
     alert_group4.personal_log_records.update(created_at=now - timezone.timedelta(minutes=2))
 
     # trigger task
-    with patch("apps.alerts.tasks.check_escalation_finished.check_personal_notifications_task") as mock_check_notif:
+    with patch(
+        "apps.alerts.tasks.check_escalation_finished.check_alert_group_personal_notifications_task"
+    ) as mock_check_notif:
         check_escalation_finished_task()
 
     for alert_group in alert_groups:
         mock_check_notif.apply_async.assert_any_call((alert_group.id,))
-        check_personal_notifications_task(alert_group.id)
+        check_alert_group_personal_notifications_task(alert_group.id)
         if alert_group == alert_group3:
             assert f"Alert group {alert_group3.id} has (1) uncompleted personal notifications" in caplog.text
         else:
             assert f"Alert group {alert_group.id} personal notifications check passed" in caplog.text
 
     mocked_send_alert_group_escalation_auditor_task_heartbeat.assert_called()
+
+    # also trigger the general personal notification checker
+    check_personal_notifications_task()
+
+    assert "personal_notifications_triggered=5 personal_notifications_completed=2" in caplog.text