+++ engine/apps/mobile_app/utils.py
@@ -5,7 +5,7 @@ import typing
 import requests
 from django.conf import settings
 from firebase_admin.exceptions import FirebaseError
-from firebase_admin.messaging import AndroidConfig, APNSConfig, APNSPayload, Message
+from firebase_admin.messaging import AndroidConfig, APNSConfig, APNSPayload, Message, UnregisteredError
 from requests import HTTPError
 from rest_framework import status
 
@@ -19,6 +19,13 @@ if typing.TYPE_CHECKING:
 
 
 MAX_RETRIES = 1 if settings.DEBUG else 10
+
+# UnregisteredError
+# App instance was unregistered from FCM. This usually means that the token used is no longer valid and a
+# new one must be used.
+#
+# In other words, this error occurs outside of our control and retrying will never fix it, therefore we should skip
+FIREBASE_ERRORS_TO_NOT_RETRY = (UnregisteredError,)
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
@@ -37,7 +44,7 @@ def _send_push_notification_to_fcm_relay(message: Message) -> requests.Response:
     return response
 
 
-def send_message_to_fcm_device(device: "FCMDevice", message: Message) -> None:
+def send_message_to_fcm_device(device: "FCMDevice", message: Message) -> bool:
     """
     https://firebase.google.com/docs/cloud-messaging/http-server-ref#interpret-downstream
     """
@@ -46,14 +53,17 @@ def send_message_to_fcm_device(device: "FCMDevice", message: Message) -> None:
 
     if isinstance(response, FirebaseError):
         logger.exception(
-            f"FCM error occured in mobile_app.utils.send_message_to_fcm_device\n"
-            f"FCMDevice info: {device}\n"
-            f"FirebaseError code: {response._code}\n"
-            f"FirebaseError cause: {response._cause}\n"
-            f"FirebaseError http_response: {response._http_response}\n"
+            f"FCM error occured in mobile_app.utils.send_message_to_fcm_device fcm_device_info={device} "
+            f"firebase_error_code={response._code} firebase_error_cause={response._cause} "
+            f"firebase_error_http_response={response._http_response}"
         )
 
+        if isinstance(response, FIREBASE_ERRORS_TO_NOT_RETRY):
+            logger.warning(f"FCM error {response} is not being retried as we explicitly do not want to retry it")
+            return False
+
         raise response
+    return True
 
 
 def send_push_notification(
@@ -88,7 +98,10 @@ def send_push_notification(
             else:
                 raise
     else:
-        send_message_to_fcm_device(device_to_notify, message)
+        succeeded = send_message_to_fcm_device(device_to_notify, message)
+        if not succeeded:
+            _error_cb()
+            return False
 
     # notification succeeded (otherwise raised exception before)
     return True